WORKAROUND going half speed
The roboteq board has been going exactly half of its commanded speed,
probably due to a mis-configuration. Here we work around it by making
this code think its max speed is half of what it is so that it will
double the speed command (which is specified as a percentage of its max
speed).

This is a temporary workaround until we fix the configuration of the
roboteq controller.